+++ parseCmdLine.h
@@ -1,0 +1,6 @@
+#include<stdlib.h>
+
+typedef int (*pCallback) (char*, char*, void*);
+
+int parseCmdLine(int argc, char* argv[], pCallback p, void* userData);
+
