+++ tests.cpp
@@ -1,4 +1,4 @@
-#include "lib.h"
+#include "parseCmdLine.h"
 #include <string>
 #include <iostream>
