+++ x64/Debug/EDA---TP1.log
@@ -1,2 +1,4 @@
-﻿  lib.cpp
+﻿  parseCmdLine.cpp
+  tests.cpp
+  Generating Code...
   EDA---TP1.vcxproj -> F:\ReposVS\EDA---TP1\x64\Debug\EDA---TP1.exe
